rename 3 zz_-files
