+++ airship_pathing.py
