+++ root_mem.py
